functions/plaid/requirements.txt
plaid-python==8.6.0
flask==2.0.2
google-cloud-firestore==2.3.4
pandas==1.3.5
numpy==1.22.0